requirements.txt
Flask==1.1.2
gunicorn==20.0.4
mysql-connector-python==8.0.21

